fix: arregla 404 en modos single-player y cuelgue de multijugador cuando no hay preguntas
